User tests + 1 error in public Tests

+++ src/main/java/com/example/controller/OrderController.java
@@ -35,7 +35,14 @@ public class OrderController {
 
     @DeleteMapping("/delete/{orderId}")
     public String deleteOrderById(@PathVariable UUID orderId){
-        orderService.deleteOrderById(orderId);
-        return "Order deleted successfully";
+
+        if(orderService.getOrderById(orderId) == null){
+            return "Order not found";
+        }
+        else {
+            orderService.deleteOrderById(orderId);
+            return "Order deleted successfully";
+        }
+
     }
 }
